airbnb-burla/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "airbnb-burla"
version = "0.1.0"
description = "Analyzing 3M Airbnb listings worldwide on Burla."
requires-python = ">=3.12"
dependencies = [
    "burla>=1.5.7",
    "pandas>=2.2",
    "pyarrow>=15",
    "numpy>=1.26",
    "requests>=2.32",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "pillow>=10",
    "anthropic>=0.34",
    "open-clip-torch>=2.24",
    "torch>=2.2",
    "torchvision>=0.17",
    "ultralytics>=8.2",
    "sentence-transformers>=3",
    "scikit-learn>=1.4",
    "langdetect>=1.0.9",
    "beautifulsoup4>=4.12",
    "lxml>=5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
